+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Option to create new panel from selection on existing panel display. (TODO)
 - Option to duplicate an existing panel. (TODO)
+- Setter element CrescendoOverride.
+
+### Fixed
+
+- Display of correct default text on setter button elements.
 
 ## [0.14.0] - 2024-07-02